CLASSES: IMAGE_TYPES-STUBI: set empty UBI volume and fix
- manage empty UBI volume for multivolume UBI
Add new parameter to allow to add empty UBI volume when creating the
multivolume UBI.
User just need to provide new volume name and size.

- clean temporary file to avoid issue while deploying to image
If we keep the temporary cfg file for empty UBI volume, we have an issue
when bitbake deploy artifacts to DEPLOYDIR_IMAGE when a different image
has been already built as files have the same name.
As we're using this cfg file as a temporary step for multi-volume cfg, we
can remove them instead of deploying them.

- manage extra_space on UBI volume for bad blocks management

Change-Id: If7ef9a66fb058cdaf0d77a22ca0be70fa05fca6f
Signed-off-by: Romuald JEANNE <romuald.jeanne@st.com>