UBOOT-STM32MP-EXTLINUX: fix boot.cmd for overlay
By default U-Boot reserve overlay sizee with 4KB padding, sometime it is
not enought and it is normal.

The caller NEED to reserved enought memory before to apply overlay
by usisng the <extrasize> parameter.

fdt resize [<extrasize>]
- Resize fdt to size + padding to 4k addr + some optional <extrasize> if needed

A correct value cannot be proposed (it is depending of the overlay size)
but at least you must reserved the size of the loaded overlay dtbo file
=> ${filesize} (this variable automatically updated after a load command)

Change-Id: Ib37f2778bd843436b0baddb67a7a4fd5c799b749
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>