TF-A-STM32MP: add file type for deploying image
When a compilation a made for TF-a with 'all' target, several binaries are generated
and have the same name on several kind of build but generated with different
compilation switch. In some case (like arm64 bits) the bl2 need to generate a
bl31 binary but without the specific option requested by the generation of official bl31,
which can generate an issue when you deploy the file because it's the binary of
last build which are used but it's not in all case the binary with compilation
option desired.
For solving this king of issue, a flag <file type> which can be : bl31 b32 bl2 fwconfig
are added on the TF_A_CONFIG to specify which kind of binary we need to export.

Change-Id: Ibd484ae2e285bf70e3e0790e020d2ea595506a3c
Signed-off-by: Christophe Priouzeau <christophe.priouzeau@foss.st.com>