CONF: MACHINE: use weak assignment for PREFERRED_PROVIDER
By its nature, the PREFERRED_PROVIDERs might be subject to change in upper
layers with a line such as:

  PREFERRED_PROVIDER_virtual/kernel = "linux-foo"

However adding the above line to local.conf to do a local experiment would
not work as users might expect because the meta-st-stm32mp layer uses the
'=' assignment, and local.conf is parsed earlier.

Relying on parsing order is usually not a good idea, thus change all the
PREFERRED_PROVIDER assignments in the machine files to using a weak
assignment, like many other silicon vendor layers do. Using the ??= allows
2 further levels of overriding, e.g. by a board/SoM manufacturer and by a
final integration layer.

Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>