GCC: make symbols are dynamically available and plugins work
The Canadian build will skip the check for rdynamic. And hence the gcc
symbols are not dynamically available and the gcc plugins won't work. The
opt_pass.*set_pass_param symbols are there but not as exported symbols
so nm -D won't see them e.g..

Based on the patch found here
https://lists.openembedded.org/g/openembedded-core/topic/rfc_build_linux_with_a_oe/90687587

Change-Id: I4860565319a786f5cc511f50cbfcabad9278ed4b
Signed-off-by: Lionel VITTE <lionel.vitte@st.com>